Puesta a punto para reports

+++ app/Observers/ReportObserver.php
@@ -18,15 +18,22 @@ class ReportObserver
 
     public function deleting(Report $report){
         $today = Carbon::today()->setTime(0, 0,0)->format('Y-m-d');
-        $path = request()->path();
         $reportDate = Carbon::parse($report->published_at)->setTime(0, 0,0);
 
-        if($reportDate->greaterThan($today)){
-            $report->forceDelete();
-        } else if($reportDate->lessThanOrEqualTo($today) && $path !== 'api/admin/reports/'.$report->id){
+        if($reportDate->lessThanOrEqualTo($today) && auth()->user()->role !== 'admin'){
             throw ValidationException::withMessages([
                 'name' => 'You cannot delete this Report'
             ]);
+        }
+    }
+
+    public function deleted(Report $report)
+    {
+        $today = Carbon::today()->setTime(0, 0,0)->format('Y-m-d');
+        $reportDate = Carbon::parse($report->published_at)->setTime(0, 0,0);
+        
+        if($reportDate->greaterThan($today) && auth()->user()->id === $report->user_id){
+            $report->forceDelete();
         }
     }
 }
